feat: add global automated moderation whitelist

+++ sophie_bot/modules/welcomesecurity/schedules/kick_unpassed_users.py
@@ -10,6 +10,7 @@ from sophie_bot.modules.restrictions.utils.restrictions import execute_restricti
 from sophie_bot.services.application import ApplicationServices
 from sophie_bot.shared.actions import RestrictionAction
 from sophie_bot.utils.feature_flags import is_enabled
+from sophie_bot.utils.global_whitelist import is_user_globally_whitelisted
 from sophie_bot.utils.logger import log
 
 
@@ -40,6 +41,10 @@ class KickUnpassedUsers:
                 "kick_unpassed_users: skipping ws_user due to missing linked user/group",
                 ws_user_tid=str(ws_user.id),
             )
+            await ws_user.delete()
+            return
+        if await is_user_globally_whitelisted(user.tid):
+            log.debug("kick_unpassed_users: removing exempt user from pending captcha", user=user.tid)
             await ws_user.delete()
             return
         if not await is_enabled(
